# openapi.yaml
openapi: 3.0.3
info:
  title: Autopilon Public API
  version: "1.1"
  description: |
    Machine-readable reference for the Autopilon Public API (/api/v1), status: beta.
    This spec covers exactly the endpoints implemented in server/publicApi/ — see
    PUBLIC_API.md for prose documentation, WEBHOOKS.md for webhook details, and
    API_SECURITY.md for the auth/tenancy/threat model.
  contact:
    url: https://github.com/beautybeespk-a11y/autopilon-mobile
servers:
  - url: https://<your-deployment>/api/v1
  - url: http://localhost:4000/api/v1
security:
  - ApiKeyAuth: []
...
paths:
...
  /integrations/{provider}/actions:
    get:
      operationId: listIntegrationActions
      summary: List curated, publicly-approved actions for a connected provider
      tags: [Integrations]
      parameters:
        - name: provider
          in: path
          required: true
          schema: { type: string }
          example: gmail
      responses:
        "200":
          description: Approved actions for this provider.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/IntegrationAction" }
        "404": { $ref: "#/components/responses/NotFound" }
  /integrations/{provider}/actions/{actionName}/execute:
    post:
      operationId: executeIntegrationAction
      summary: Execute one approved integration action
      tags: [Integrations]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - name: provider
          in: path
          required: true
          schema: { type: string }
          example: gmail
        - name: actionName
          in: path
          required: true
          schema: { type: string }
          example: gmail.list_emails
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [agentId]
              properties:
                agentId:
                  type: string
                  description: An agent in this org with the matching skill enabled — the action runs through that agent's own tool permissions.
                parameters:
                  type: object
                  additionalProperties: true
      responses:
        "200":
          description: Execution result (completed, failed, or awaiting_confirmation - all normal 200 outcomes, not HTTP errors).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IntegrationActionResult" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "403": { $ref: "#/components/responses/InsufficientScope" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: The API key creator's active organization does not match this organization's connected integration (INTEGRATION_CONTEXT_MISMATCH).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: "ap_live_<key>"
      description: |
        Bearer API key issued from an organization's Developer Console.
        Never a session cookie. See API_SECURITY.md.
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: |
        Optional. A retried request with the same key and identical body replays
        the original response instead of running again. See PUBLIC_API.md's
        Idempotency section for full semantics (conflict/in-progress/expiry).
      schema: { type: string, maxLength: 255 }
  responses:
...
    InsufficientScope:
      description: The API key lacks the scope this route requires.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    NotFound:
      description: No such resource in this organization.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    InvalidRequest:
      description: Missing/malformed request body or query parameter.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: "RESOURCE_NOT_FOUND" }
            message: { type: string }
            request_id: { type: string, nullable: true }
...
    IntegrationAction:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        parameters: { type: object, additionalProperties: true }
        requiresConfirmation: { type: boolean }
    IntegrationActionResult:
      type: object
      properties:
        executionId: { type: string }
        status: { type: string, enum: [completed, failed, awaiting_confirmation] }
        result: { type: object, additionalProperties: true, nullable: true }
        error: { type: string, nullable: true }
        confirmationId: { type: string, nullable: true }
        reason: { type: string, nullable: true }